+++ src/main/java/core/model/player/Player.java
@@ -52,7 +52,9 @@ public class Player {
     /**
      * Name
      */
-    private String m_sName = "";
+    private String m_sFirstName = "";
+    private String m_sNickName = "";
+    private String m_sLastName = "";
 
     /**
      * TeamInfo Smilie Filename
@@ -350,7 +352,9 @@ public class Player {
         // Separate first, nick and last names are available. Utilize them?
 
         m_iSpielerID = Integer.parseInt(properties.getProperty("id", "0"));
-        m_sName = properties.getProperty("name", "");
+        m_sFirstName = properties.getProperty("firstname", "");
+        m_sNickName = properties.getProperty("nickname", "");
+        m_sLastName = properties.getProperty("lastname", "");
         m_iAlter = Integer.parseInt(properties.getProperty("ald", "0"));
         m_iAgeDays = Integer.parseInt(properties.getProperty("agedays", "0"));
         m_iKondition = Integer.parseInt(properties.getProperty("uth", "0"));
@@ -467,8 +471,12 @@ public class Player {
     }
 
     /**
-     * liefert das Datum des letzen LevelAufstiegs für den angeforderten Skill Vector filled with
-     * object[] [0] = Time der Änderung [1] = Boolean: false=Keine Änderung gefunden
+     * gives information of skill ups
+     * returns vector of
+     * object[]
+     *      [0] = date of skill up
+     *      [1] = Boolean: false=no skill up found
+     *      [2] = skill value
      */
     public Vector<Object[]> getAllLevelUp(int skill) {
         return DBManager.instance().getAllLevelUp(skill, m_iSpielerID);
@@ -920,6 +928,10 @@ public class Player {
         m_clhrfDate = timestamp;
     }
 
+    public void setHrfDate() {
+        Date now = new Date();
+        setHrfDate(new Timestamp(now.getTime()));
+    }
 
     /**
      * calculate the contribution for the ideal position
@@ -1174,46 +1186,57 @@ public class Player {
 //		return HOVerwaltung.instance().getModel().getEPV().getPrice(data);
 //    }
 
-    /**
-     * Setter for property m_sName.
-     *
-     * @param m_sName New value of property m_sName.
-     */
-    public void setName(java.lang.String m_sName) {
-        this.m_sName = m_sName;
+
+    public void setFirstName(java.lang.String m_sName) {
+        this.m_sFirstName = m_sName;
     }
 
-    /**
-     * Getter for property m_sName.
-     *
-     * @return Value of property m_sName.
-     */
-    public java.lang.String getName() {
-        return DBManager.deleteEscapeSequences(m_sName);
+    public java.lang.String getFirstName() {
+        return DBManager.deleteEscapeSequences(m_sFirstName);
+    }
+
+    public void setNickName(java.lang.String m_sName) {
+        this.m_sNickName = m_sName;
+    }
+
+    public java.lang.String getNickName() {
+        return DBManager.deleteEscapeSequences(m_sNickName);
+    }
+
+    public void setLastName(java.lang.String m_sName) {
+        this.m_sLastName = m_sName;
     }
 
     public java.lang.String getLastName() {
-        String lastName = this.getName();
-        return lastName.substring(lastName.lastIndexOf(" ") + 1);
+        return DBManager.deleteEscapeSequences(m_sLastName);
     }
+
+
 
     /**
      * Getter for shortName
-     *
-     * @return returns the fist letter of the first Name + a "." and the last name
-     * eg: James Bond = J. Bond, unless the string is already empty — in which
-     * case it is deemed short enough, and returned.
+     * eg: James Bond = J. Bond
+     * Nickname are ignored
      */
     public String getShortName() {
-        String fullName = getName();
 
-        if (StringUtils.isEmpty(fullName)) {
-            return fullName;
-        } else {
-            String initial = fullName.substring(0, 1);
-            String lastName = getLastName();
-            return initial + ". " + lastName;
+        if (getFirstName().isEmpty())
+        {
+            return getLastName();
         }
+        return getFirstName().substring(0, 1) + ". " + getLastName();
+
+        }
+
+
+    public java.lang.String getFullName() {
+
+        if (getNickName().isEmpty())
+        {
+            return getFirstName() + " " +getLastName();
+        }
+
+        return getFirstName() + " '" + getNickName() + "' " +getLastName();
     }
 
     /**
@@ -2259,6 +2282,7 @@ public class Player {
     public void setSubExperience( double experience){
         this.subExperience = experience;
     }
+
 }
 
 class PositionContribute {