+++ src/main/java/module/lineup/substitution/positionchooser/PositionChooser.java
@@ -190,7 +190,7 @@ public class PositionChooser extends JPanel {
 			this.player = player;
 			if (isOccupied()) {
 				setBackground(COLOR_POS_OCCUPIED);
-				setToolTipText(player.getName());
+				setToolTipText(player.getFullName());
 			} else {
 				setBackground(COLOR_POS_DEFAULT);
 				setToolTipText(null);