+++ src/main/java/module/playerOverview/SpielerTrainingsSimulatorPanel.java
@@ -28,6 +28,8 @@ import java.awt.event.ActionListener;
 import java.awt.event.FocusEvent;
 import java.awt.event.FocusListener;
 import java.awt.event.ItemListener;
+import java.sql.Date;
+import java.sql.Timestamp;
 
 import javax.swing.JButton;
 import javax.swing.JCheckBox;
@@ -193,10 +195,11 @@ final class SpielerTrainingsSimulatorPanel extends ImagePanel
     public final void actionPerformed(ActionEvent e) {
         if (e.getSource().equals(m_jbAddTempSpieler)) {
             final Player tempPlayer = new Player();
+            tempPlayer.setHrfDate();
             tempPlayer.setNationalitaet(HOVerwaltung.instance().getModel().getBasics().getLand());
             tempPlayer.setSpielerID(module.transfer.scout.TransferEingabePanel
                     .getNextTempSpielerID());
-            tempPlayer.setName("Temp " + Math.abs(1000 + tempPlayer.getSpielerID()));
+            tempPlayer.setLastName("Temp " + Math.abs(1000 + tempPlayer.getSpielerID()));
             tempPlayer.setAlter(getAge());
             tempPlayer.setAgeDays(getAgeDays());
             tempPlayer.setErfahrung(((CBItem) m_jcbErfahrung.getSelectedItem()).getId());
@@ -241,7 +244,7 @@ final class SpielerTrainingsSimulatorPanel extends ImagePanel
     }
 
     private void setCBs() {
-        m_jlName.setText(m_clPlayer.getName());
+        m_jlName.setText(m_clPlayer.getFullName());
         jtfAge.setText(m_clPlayer.getAlter() + "." + m_clPlayer.getAgeDays());
         Helper.markierenComboBox(m_jcbForm, m_clPlayer.getForm());
         Helper.markierenComboBox(m_jcbErfahrung, m_clPlayer.getErfahrung());